backend/pyproject.toml
[project]
name = "fontegest-api"
version = "0.1.0"
description = "FonteGest Condomínios — API Backend"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.10.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0,<4.1",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "sqladmin>=0.20.0",
    "itsdangerous>=2.2.0",
    "openpyxl>=3.1.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.41.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.20.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
